Extract user creation
