+++ src/services/sync/strategies/TSP/TSPSchoolSyncer.js
@@ -7,9 +7,8 @@ const {
 	TspApi,
 	config: TSP_CONFIG,
 	ENTITY_SOURCE, SOURCE_ID_ATTRIBUTE,
-	getUsername, getEmail,
+	createUserAndAccount,
 } = require('./TSP');
-const accountModel = require('../../../account/model');
 
 const SYNCER_TARGET = 'tsp-school';
 
@@ -249,15 +248,19 @@ class TSPSchoolSyncer extends mix(Syncer).with(ClassImporter) {
 		try {
 			const sourceOptions = {};
 			sourceOptions[SOURCE_ID_ATTRIBUTE] = tspTeacher.lehrerUid;
-			const teacher = await this.createUserAndAccount({
-				namePrefix: tspTeacher.lehrerTitel,
-				firstName: tspTeacher.lehrerVorname,
-				lastName: tspTeacher.lehrerNachname,
-				schoolId: school._id,
-				source: ENTITY_SOURCE,
-				sourceOptions,
-			},
-			'teacher', systemId);
+			const teacher = await createUserAndAccount(
+				this.app,
+				{
+					namePrefix: tspTeacher.lehrerTitel,
+					firstName: tspTeacher.lehrerVorname,
+					lastName: tspTeacher.lehrerNachname,
+					schoolId: school._id,
+					source: ENTITY_SOURCE,
+					sourceOptions,
+				},
+				'teacher',
+				systemId,
+			);
 			this.stats.users.teachers.created += 1;
 			return teacher;
 		} catch (err) {
@@ -334,14 +337,18 @@ class TSPSchoolSyncer extends mix(Syncer).with(ClassImporter) {
 		try {
 			const sourceOptions = {};
 			sourceOptions[SOURCE_ID_ATTRIBUTE] = tspStudent.schuelerUid;
-			const student = await this.createUserAndAccount({
-				firstName: tspStudent.schuelerVorname,
-				lastName: tspStudent.schuelerNachname,
-				schoolId: school._id,
-				source: ENTITY_SOURCE,
-				sourceOptions,
-			},
-			'student', systemId);
+			const student = await createUserAndAccount(
+				this.app,
+				{
+					firstName: tspStudent.schuelerVorname,
+					lastName: tspStudent.schuelerNachname,
+					schoolId: school._id,
+					source: ENTITY_SOURCE,
+					sourceOptions,
+				},
+				'student',
+				systemId,
+			);
 			this.stats.users.students.created += 1;
 			return student;
 		} catch (err) {
@@ -355,37 +362,6 @@ class TSPSchoolSyncer extends mix(Syncer).with(ClassImporter) {
 			});
 			return null;
 		}
-	}
-
-	/**
-	 * Registers a user and creates an account
-	 * @param {Object} userOptions options to be provided to the user service
-	 * @param {Array<String>} roles the user's roles
-	 * @param {System} systemId the user's login system
-	 * @returns {User} the user object
-	 * @async
-	 */
-	async createUserAndAccount(userOptions, roles, systemId) {
-		const username = getUsername(userOptions);
-		const email = getEmail(userOptions);
-		const { pin } = await this.app.service('registrationPins').create({
-			email,
-			verified: true,
-			silent: true,
-		});
-		const user = await this.app.service('users').create({
-			...userOptions,
-			pin,
-			email,
-			roles,
-		});
-		await accountModel.create({
-			userId: user._id,
-			username,
-			systemId,
-			activated: true,
-		});
-		return user;
 	}
 
 	/**
